Gallery Block: Select and Remove an image
