+++ blocks/library/gallery/editor.scss
@@ -9,18 +9,23 @@
 
 .blocks-gallery-image {
 	position: relative;
+
+	&.is-selected {
+		outline: 4px solid $blue-medium-500;
+		outline-offset: -4px;
+	}
+}
+
+.blocks-gallery-image__inline-menu {
+	position: absolute;
+	top: 8px;
+	right: 8px;
+	border: 1px solid $light-gray-500;
+	background-color: $white;
+	display: inline-flex;
+	z-index: z-index( '.blocks-gallery-image__inline-menu' );
 }
 
 .blocks-gallery-image__remove {
-	position: absolute;
-	top: 5px;
-	right: 5px;
 	padding: 0;
-	border: 1px solid $light-gray-900;
-	background: $white;
-	opacity: 0.8;
-
-	&:hover {
-		opacity: 1;
-	}
 }
